made file public
